+++ Common/src/main/java/com/clinics/common/DTO/response/outer/DoctorResponseDTO.java
@@ -3,7 +3,6 @@ package com.clinics.common.DTO.response.outer;
 import lombok.*;
 
 import java.util.Collection;
-import java.util.List;
 import java.util.UUID;
 
 @Data
