+++ Doctors-MSSC/src/main/java/com/clinics/doctors/ui/client/DoctorMedicalUnitController.java
@@ -1,7 +1,7 @@
 package com.clinics.doctors.ui.client;
 
 import com.clinics.common.DTO.response.outer.MedicalUnitResponseDTO;
-import com.clinics.doctors.ui.service.DoctorMedicalUnitClient;
+import com.clinics.doctors.ui.service.JPAimpl.DoctorMedicalUnitClientImpl;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -15,16 +15,16 @@ import java.util.UUID;
 public class DoctorMedicalUnitController {
 
 
-	final private DoctorMedicalUnitClient doctorMedicalUnitClient;
+	final private DoctorMedicalUnitClientImpl doctorMedicalUnitClientImpl;
 
 	public DoctorMedicalUnitController(
-			DoctorMedicalUnitClient doctorMedicalUnitClient) {
-		this.doctorMedicalUnitClient = doctorMedicalUnitClient;
+			DoctorMedicalUnitClientImpl doctorMedicalUnitClientImpl) {
+		this.doctorMedicalUnitClientImpl = doctorMedicalUnitClientImpl;
 	}
 
 	@GetMapping
 	public ResponseEntity<List<MedicalUnitResponseDTO>> getAllDoctorMedicalUnites(@PathVariable UUID doctorUUID){
-		return ResponseEntity.ok().body(doctorMedicalUnitClient.getAll(doctorUUID));
+		return ResponseEntity.ok().body(doctorMedicalUnitClientImpl.getAll(doctorUUID));
 	}
 
 	@GetMapping(value = "/{medicalUnitUUID}")
@@ -32,21 +32,21 @@ public class DoctorMedicalUnitController {
 			@PathVariable UUID doctorUUID,
 			@PathVariable UUID medicalUnitUUID
 	){
-		return ResponseEntity.ok().body(doctorMedicalUnitClient.get(doctorUUID, medicalUnitUUID));
+		return ResponseEntity.ok().body(doctorMedicalUnitClientImpl.get(doctorUUID, medicalUnitUUID));
 	}
 
 	@PostMapping(value = "/{medicalUnitUUID}")
 	public ResponseEntity<MedicalUnitResponseDTO> addMedicalUniteIntoDoctor(
 			@PathVariable UUID medicalUnitUUID,
 			@PathVariable UUID doctorUUID) {
-		return ResponseEntity.status(HttpStatus.CREATED).body(doctorMedicalUnitClient.save(medicalUnitUUID, doctorUUID));
+		return ResponseEntity.status(HttpStatus.CREATED).body(doctorMedicalUnitClientImpl.save(medicalUnitUUID, doctorUUID));
 	}
 
 	@DeleteMapping(value = "/{medicalUnitUUID}")
 	public ResponseEntity<Void> removeMedicalUniteFromDoctor(
 			@PathVariable UUID medicalUnitUUID,
 			@PathVariable UUID doctorUUID) {
-		doctorMedicalUnitClient.delete(doctorUUID, medicalUnitUUID);
+		doctorMedicalUnitClientImpl.delete(doctorUUID, medicalUnitUUID);
 		return ResponseEntity.ok().build();
 	}
 }
