+++ Doctors-MSSC/src/main/java/com/clinics/doctors/ui/controller/DoctorAppointmentController.java
@@ -2,7 +2,7 @@ package com.clinics.doctors.ui.controller;
 
 import com.clinics.common.DTO.request.outer.doctor.AppointmentDTO;
 import com.clinics.common.DTO.response.outer.AppointmentResponseDTO;
-import com.clinics.doctors.ui.service.AppointmentService;
+import com.clinics.doctors.ui.service.JPAimpl.AppointmentServiceImpl;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -16,18 +16,18 @@ import java.util.UUID;
 @RequestMapping(value = "/doctors/{doctorUUID}/calendars/{calendarUUID}/appointments")
 public class DoctorAppointmentController {
 
-	final private AppointmentService appointmentService;
+	final private AppointmentServiceImpl appointmentServiceImpl;
 
 	public DoctorAppointmentController(
-			AppointmentService appointmentService) {
-		this.appointmentService = appointmentService;
+			AppointmentServiceImpl appointmentServiceImpl) {
+		this.appointmentServiceImpl = appointmentServiceImpl;
 	}
 
 	@GetMapping
 	public ResponseEntity<List<AppointmentResponseDTO>> getAllCalendarAppointments(
 			@PathVariable UUID doctorUUID,
 			@PathVariable UUID calendarUUID){
-		return ResponseEntity.ok().body(appointmentService.getAllDoctorAppointments(doctorUUID, calendarUUID));
+		return ResponseEntity.ok().body(appointmentServiceImpl.getAllDoctorAppointments(doctorUUID, calendarUUID));
 	}
 
 	@GetMapping(value = "/{appointmentUUID}")
@@ -35,7 +35,7 @@ public class DoctorAppointmentController {
 			@PathVariable UUID doctorUUID,
 			@PathVariable UUID calendarUUID,
 			@PathVariable UUID appointmentUUID){
-		return ResponseEntity.ok().body(appointmentService.getDoctorAppointment(doctorUUID, calendarUUID, appointmentUUID));
+		return ResponseEntity.ok().body(appointmentServiceImpl.getDoctorAppointment(doctorUUID, calendarUUID, appointmentUUID));
 	}
 
 	@PostMapping
@@ -43,7 +43,7 @@ public class DoctorAppointmentController {
 			@Valid @RequestBody AppointmentDTO appointmentDTO,
 			@PathVariable UUID doctorUUID,
 			@PathVariable UUID calendarUUID){
-		return ResponseEntity.status(HttpStatus.CREATED).body(appointmentService.saveAppointment(doctorUUID, calendarUUID, appointmentDTO));
+		return ResponseEntity.status(HttpStatus.CREATED).body(appointmentServiceImpl.saveAppointment(doctorUUID, calendarUUID, appointmentDTO));
 	}
 
 	@PostMapping(value = "/multiple")
@@ -51,7 +51,7 @@ public class DoctorAppointmentController {
 			@Valid @RequestBody List<AppointmentDTO> addEditAppointmentsDTO,
 			@PathVariable UUID doctorUUID,
 			@PathVariable UUID calendarUUID){
-		return ResponseEntity.status(HttpStatus.CREATED).body(appointmentService.saveAppointments(doctorUUID, calendarUUID, addEditAppointmentsDTO));
+		return ResponseEntity.status(HttpStatus.CREATED).body(appointmentServiceImpl.saveAppointments(doctorUUID, calendarUUID, addEditAppointmentsDTO));
 	}
 
 	@PatchMapping(value = "/{appointmentUUID}")
@@ -60,7 +60,7 @@ public class DoctorAppointmentController {
 			@PathVariable UUID calendarUUID,
 			@PathVariable UUID appointmentUUID,
 			@Valid @RequestBody AppointmentDTO appointmentDTO){
-		appointmentService.editAppointment(doctorUUID, calendarUUID, appointmentUUID, appointmentDTO);
+		appointmentServiceImpl.editAppointment(doctorUUID, calendarUUID, appointmentUUID, appointmentDTO);
 		return ResponseEntity.ok().build();
 	}
 
@@ -69,7 +69,7 @@ public class DoctorAppointmentController {
 			@PathVariable UUID doctorUUID,
 			@PathVariable UUID calendarUUID,
 			@PathVariable UUID appointmentUUID){
-		appointmentService.deleteAppointment(doctorUUID, calendarUUID, appointmentUUID);
+		appointmentServiceImpl.deleteAppointment(doctorUUID, calendarUUID, appointmentUUID);
 		return ResponseEntity.ok().build();
 	}
 }
